chore(clerk-js): Move all SignUp tests in the __tests__ directory
