+++ packages/clerk-js/src/ui/components/SignUp/__tests__/util.test.ts
@@ -1,6 +1,6 @@
 import { SignUpResource } from '@clerk/types';
 
-import { completeSignUpFlow } from './util';
+import { completeSignUpFlow } from '../util';
 
 const mockHandleComplete = jest.fn();
 const mockNavigate = jest.fn();
